feat: add async iterable generator method generateFrames
The generator yields an iterable of output strings (at different
animation frames) for a given time duration.

The core idea is to provide a way to render the animation in any
environment (browser or node). While `startAnimation()` method directly
prints the ASCII animation in console, the `generateFrames()` method
generates the strings to be used anywhere accordingly.